+++ cadence-macros/src/state.rs
@@ -8,31 +8,98 @@
 // option. This file may not be copied, modified, or distributed
 // except according to those terms.
 
-use cadence::prelude::*;
+use cadence::StatsdClient;
+use std::cell::UnsafeCell;
 use std::error::Error;
 use std::fmt::{self, Display, Formatter};
-use std::panic::RefUnwindSafe;
-use std::sync::{Arc, Once};
+use std::sync::atomic::{AtomicUsize, Ordering};
+use std::sync::Arc;
 
-type GlobalClient = dyn MetricClient + Send + Sync + RefUnwindSafe + 'static;
+const UNSET: usize = 0;
+const LOADING: usize = 1;
+const COMPLETE: usize = 2;
 
-static GLOBAL_INIT: Once = Once::new();
-static mut GLOBAL_DEFAULT: Option<Arc<GlobalClient>> = None;
+/// Global default StatsdClient to be used by macros
+static HOLDER: SingletonHolder<StatsdClient> = SingletonHolder::new();
 
-/// Error indicating that a global default `MetricClient` was not set
+/// Holder to allow global reads of a value from multiple threads while
+/// allowing the value to be written (set) a single time.
+///
+/// This type is public to allow it to be used in integration tests for
+/// this crate but it is not part of the public API and may change at any
+/// time.
+#[doc(hidden)]
+#[derive(Debug)]
+pub struct SingletonHolder<T> {
+    value: UnsafeCell<Option<Arc<T>>>,
+    state: AtomicUsize,
+}
+
+impl<T> SingletonHolder<T> {
+    /// Create a new empty holder
+    pub const fn new() -> Self {
+        SingletonHolder {
+            value: UnsafeCell::new(None),
+            state: AtomicUsize::new(UNSET),
+        }
+    }
+}
+
+impl<T> SingletonHolder<T> {
+    /// Get a pointer to the contained value if set, None otherwise
+    pub fn get(&self) -> Option<Arc<T>> {
+        if !self.is_set() {
+            return None;
+        }
+
+        // SAFETY: We've ensured that the state is "complete" and the
+        // set method has completed and set a value for the UnsafeCell.
+        unsafe { &*self.value.get() }.clone()
+    }
+
+    pub fn is_set(&self) -> bool {
+        COMPLETE == self.state.load(Ordering::Acquire)
+    }
+
+    /// Set the value if it has not already been set, otherwise this is a no-op
+    #[allow(deprecated)]
+    pub fn set(&self, val: T) {
+        // compare_and_swap is deprecated in 1.50 but we target back to 1.36
+        if UNSET != self.state.compare_and_swap(UNSET, LOADING, Ordering::SeqCst) {
+            return;
+        }
+
+        // SAFETY: There are no readers at this point since we've guaranteed the
+        // state could not have been "complete". There are no other writers since
+        // we've ensured that the state was previously "unset" and we've been able
+        // to compare-and-swap it to "loading".
+        let ptr = self.value.get();
+        unsafe {
+            *ptr = Some(Arc::new(val));
+        }
+
+        self.state.store(COMPLETE, Ordering::Release);
+    }
+}
+
+unsafe impl<T: Send> Send for SingletonHolder<T> {}
+
+unsafe impl<T: Sync> Sync for SingletonHolder<T> {}
+
+/// Error indicating that a global default `StatsdClient` was not set
 /// when a call to `get_global_default` was made.
 #[derive(Debug)]
 pub struct GlobalDefaultNotSet;
 
 impl Display for GlobalDefaultNotSet {
     fn fmt(&self, f: &mut Formatter<'_>) -> fmt::Result {
-        Display::fmt("global default MetricClient instance not set", f)
+        Display::fmt("global default StatsdClient instance not set", f)
     }
 }
 
 impl Error for GlobalDefaultNotSet {}
 
-/// Set the global default `MetricClient` instance
+/// Set the global default `StatsdClient` instance
 ///
 /// If the global default client has already been set, this method does nothing.
 ///
@@ -44,18 +111,11 @@ impl Error for GlobalDefaultNotSet {}
 ///
 /// cadence_macros::set_global_default(client);
 /// ```
-pub fn set_global_default<T>(client: T)
-where
-    T: MetricClient + Send + Sync + RefUnwindSafe + 'static,
-{
-    GLOBAL_INIT.call_once(move || {
-        unsafe {
-            GLOBAL_DEFAULT = Some(Arc::new(client));
-        };
-    });
+pub fn set_global_default(client: StatsdClient) {
+    HOLDER.set(client);
 }
 
-/// Get a reference to the global default `MetricClient` instance
+/// Get a reference to the global default `StatsdClient` instance
 ///
 /// # Errors
 ///
@@ -76,11 +136,11 @@ where
 /// let global_client = cadence_macros::get_global_default();
 /// assert!(global_client.is_ok());
 /// ```
-pub fn get_global_default() -> Result<Arc<GlobalClient>, GlobalDefaultNotSet> {
-    unsafe { GLOBAL_DEFAULT.clone() }.ok_or(GlobalDefaultNotSet)
+pub fn get_global_default() -> Result<Arc<StatsdClient>, GlobalDefaultNotSet> {
+    HOLDER.get().ok_or(GlobalDefaultNotSet)
 }
 
-/// Return true if the global default `MetricClient` is set, false otherwise
+/// Return true if the global default `StatsdClient` is set, false otherwise
 ///
 /// # Example
 ///
@@ -95,6 +155,5 @@ pub fn get_global_default() -> Result<Arc<GlobalClient>, GlobalDefaultNotSet> {
 /// assert!(cadence_macros::is_global_default_set());
 /// ```
 pub fn is_global_default_set() -> bool {
-    // NOTE: not using Once::is_completed() here since it's rust 1.43+
-    unsafe { &GLOBAL_DEFAULT }.is_some()
+    HOLDER.is_set()
 }